Heavily refactor. Add icons to manifest. Handlilng license. Initial handling iAP.
Begin iAP flow needs to be initiated a call of API from content script
thus we need background script to send message to content script.
Unless window will be automatically closed suddenly after we click on
extension icon.

Verfiying license code is there now.

But the consequence of limited functionality for the extension is not
there yet.